+++ CHANGES.rst
@@ -1,6 +1,8 @@
 0.12.2 (unreleased)
 ===================
 
+- Compatibility with pytest 7.4 w.r.t. norecursedirs handling. [#201]
+
 - Respect ``--doctest-continue-on-failure`` flag. [#197]
 
 - Report doctests raising skip exceptions as skipped. [#196]